Isolate context menus for root window
